+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Client::runIterate` method (#45)
 - `ReadValueId` wrapper class (#45)
 - Example `client_subscription.cpp` (#45)
+- `ByteString::fromBase64` and `ByteString::toBase64` methods
 
 ### Changed